Input Checker works completely, normitte errors fixed

+++ libft_lib/libft.h
@@ -6,7 +6,7 @@
 /*   By: nildruon <nildruon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 16:21:27 by nildruon          #+#    #+#             */
-/*   Updated: 2026/02/03 17:28:57 by nildruon         ###   ########.fr       */
+/*   Updated: 2026/02/04 15:23:16 by nildruon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,16 +17,15 @@
 # include <stdlib.h>
 # include <stdint.h>
 # include <stdarg.h>
-#include <stdio.h>
 //Libft
 typedef struct s_list
 {
-void *content;
-struct s_list *next;
-} t_list;
+	void			*content;
+	struct s_list	*next;
+}	t_list;
 
 int			ft_atoi(const char *nptr);
-int			ft_atol(const char *nptr);
+long		ft_atol(const char *nptr);
 void		ft_bzero(void *s, size_t n);
 void		*ft_calloc(size_t nmemb, size_t size);
 int			ft_isalnum(int c);
@@ -60,17 +59,17 @@ char		*ft_strtrim(char const *s1, char const *set);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
-t_list *ft_lstnew(void *content);
-void ft_lstadd_front(t_list **lst, t_list *new);
-int ft_lstsize(t_list *lst);
-t_list *ft_lstlast(t_list *lst);
-void ft_lstadd_back(t_list **lst, t_list *new);
-void ft_lstdelone(t_list *lst, void (*del)(void*));
-void ft_lstclear(t_list **lst, void (*del)(void*));
+t_list		*ft_lstnew(void *content);
+void		ft_lstadd_front(t_list **lst, t_list *new);
+int			ft_lstsize(t_list *lst);
+t_list		*ft_lstlast(t_list *lst);
+void		ft_lstadd_back(t_list **lst, t_list *new);
+void		ft_lstdelone(t_list *lst, void (*del)(void*));
+void		ft_lstclear(t_list **lst, void (*del)(void*));
 //Printf
-int	print_hex(char c, uintptr_t nb, int len);
-int	print_pointer(void *ptr);
-int	putnbr_u_int(unsigned int n, int len);
-int	ft_printf(const char *format, ...);
+int			print_hex(char c, uintptr_t nb, int len);
+int			print_pointer(void *ptr);
+int			putnbr_u_int(unsigned int n, int len);
+int			ft_printf(const char *format, ...);
 
 #endif
